perf(web): split constant files to improve web performance
